argmin_observer_spectator

Struct SpectatorBuilder

pub struct SpectatorBuilder { /* private fields */ }

Builder for the Spectator observer

Example

use argmin_observer_spectator::SpectatorBuilder;

let spectator = SpectatorBuilder::new()
    // Optional: Name the optimization run
    // Default: random UUID.
    .with_name("optimization_run_1")
    // Optional, defaults to 127.0.0.1
    .with_host("127.0.0.1")
    // Optional, defaults to 5498
    .with_port(5498)
    // Choose which metrics should automatically be selected.
    // If omitted, all metrics will be selected.
    .select(&["cost", "best_cost"])
    // Build Spectator observer
    .build();

Implementations

impl SpectatorBuilder

pub fn new() -> Self

Creates a new SpectatorBuilder

pub fn with_name<T: AsRef<str>>(self, name: T) -> Self

Set a name the optimization run will be identified with

Defaults to a random UUID.

Example

let builder = SpectatorBuilder::new().with_name("optimization_run_1");

pub fn with_host<T: AsRef<str>>(self, host: T) -> Self

Set the host argmin spectator is running on.

Defaults to 127.0.0.1.

Example

let builder = SpectatorBuilder::new().with_host("192.168.0.1");

pub fn with_port(self, port: u16) -> Self

Set the port Spectator is running on.

Defaults to 5498.

Example

let builder = SpectatorBuilder::new().with_port(1234);

pub fn with_channel_capacity(self, capacity: usize) -> Self

Set the channel capacity

A channel is used to queue messages for sending to Spectator. If the channel capacity is reached backpressure will be applied, effectively blocking the optimization. Defaults to 10000. Decrease this value in case memory consumption is too high and increase the value in case blocking causes negative effects.

Example

let builder = SpectatorBuilder::new().with_channel_capacity(1000);

pub fn select<T: AsRef<str>>(self, metrics: &[T]) -> Self

Define which metrics will be selected in Spectator by default

If none are set, all metrics will be selected and shown. Providing zero or more metrics via select disables all apart from the provided ones. Note that independent of this setting, all data will be sent, and metrics can be selected and deselected via the Spectator GUI.

Example

let builder = SpectatorBuilder::new().select(&["cost", "best_cost"]);

pub fn name(&self) -> &String

Returns the name of the optimization run

Example

let name = builder.name();

pub fn host(&self) -> &String

Returns the host this observer will connect to

Example

let host = builder.host();

pub fn port(&self) -> u16

Returns the port this observer will connect to

Example

let port = builder.port();

pub fn channel_capacity(&self) -> usize

Returns the channel capacity

Example

let capacity = builder.channel_capacity();

pub fn selected(&self) -> &HashSet<String>

Returns the selected metrics

Example

let selected = builder.selected();

pub fn build(self) -> Spectator

Build a Spectator instance from the builder

This initiates the connection to the Spectator instance.

Example

let spectator = SpectatorBuilder::new().build();